set parseUrl to false to ensure a server instance
otherwise, if the couchdb server is not at the root path, e.g. http://localhost/couchserver/database/, then couch will be initialized as a nano instance for the „couchserver“ db on the server localhost.

This will fail on index.js:12, because a db instance does not have a request function.